Fix import of type: import A.B
Before we failed to classify imports of style A.B as first/second/third
party. This fix allows us to compare just 'A' to the lists instead of
looking for 'A.B' like before.

Heavily refactored get_paths as part of the above change.

Changed multiline imports with parenthesis to always have a trailing
comma.